Move old implementation to legacy
