+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # CHANGELOG
 
+## Next Release
+
+1. Move old starter code to a `legacy` folder so it can still be used.  (Issue #56)
+
 ## 1.5.0 (Feb 19, 2019)
 
 1. Report an exit status of 1 if DB Preparation or RSpec fails.  (Issue #51)
