+++ lib/rspec_starter/legacy/core_ext/string.rb
