+++ lib/rspec_starter/legacy/help.rb
