+++ lib/rspec_starter/legacy/runner.rb
