+++ lib/rspec_starter/legacy/steps/invoke_rspec_step.rb
