+++ lib/rspec_starter/legacy/steps/prepare_database_step.rb
