+++ lib/rspec_starter/legacy/steps/remove_tmp_folder_step.rb
