+++ lib/rspec_starter/legacy/steps/verify_xvfb_step.rb
