+++ lib/rspec_starter/legacy/which.rb
